# core/identity/contracts/auth.openapi.yaml
openapi: 3.1.0
info:
  title: BThwani Core Identity API
  version: 0.3.0
  description: Authentication, rotating sessions, actor identity, roles, permissions, and internal actor provisioning.
...
servers:
  - url: http://localhost:58082
    description: Local identity runtime
paths:
...
  /auth/session:
    get:
      operationId: getAuthSession
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated actor session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActorIdentity"
        "401":
          $ref: "#/components/responses/Unauthenticated"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    serviceToken:
      type: http
      scheme: bearer
      description: Workforce service token (IDENTITY_WORKFORCE_SERVICE_TOKEN) for internal actor APIs.
...
  responses:
...
    Unauthenticated:
      description: Authentication failed or the session expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    ActorIdentity:
      type: object
      required:
        - subject
        - tenantId
        - roles
        - permissions
        - authState
        - surfaceAccess
        - serviceAccess
        - sessionId
        - expiresAt
      properties:
        subject:
          type: string
          minLength: 1
        tenantId:
          type: string
          minLength: 1
        roles:
          type: array
          minItems: 1
          items:
            type: string
            enum: [client, partner, captain, field, operator, system]
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        authState:
          const: authenticated
        surfaceAccess:
          type: object
          additionalProperties:
            type: boolean
        serviceAccess:
          type: object
          additionalProperties:
            type: boolean
        sessionId:
          type: string
        expiresAt:
          type: string
          format: date-time
    Permission:
      type: object
      additionalProperties: false
      required: [service, surface, action, scope]
      properties:
        service:
          type: string
          enum: [dsh, wlt, knz, arb, amn, esf, mrf, snd, kwd, core]
        surface:
          type: string
          enum: [app-client, app-partner, app-captain, app-field, control-panel, webapp, website, system]
        action:
          type: string
          minLength: 1
        scope:
          type: string
          enum: [own, assigned, store, zone, all, none]
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string